MTF-calculation updated
After a discussion with Goran we found that the sigma was much too high to actually 'see' an MTF for the gauss-filtered responses.
Updated the code to show two different cases with smaller sigma.
Now desperately needs a way to put everything into a nice plot, and not such a copy-paste-kludge...