MTF.py is now much prettier and included noise
Once you figure out the specifics of Python dictionaries, you can get rid of looots of redundant code. Which is nice.
I've done that for MTF.py, now the plotting is done in 20 lines instead of 60 (for two datasets more :)
The two additional datasets include some noise, so we can start to understand the MTF better.