ReadCamera.py is now more clever and less buggy
Updated to look for the camera an not dumbly assume it's at /dev/video0.
The display of the current feed is only done in small resolution, since
large ones often overwhelmed the PC.
Images are now saved in distinct directories, so we don't overwrite
anything.
Implemented little timing function, but the timing is always between 20
and 30 images per second, so it's probably more based on ffmpeg than on
the exposure time...