(More) correct way of setting frame rate
During the experiments we saw that we need to correctly set the frame
rate. Thus introducing new option for setting the frame rate and a
better way to generate the name of the output directory (without
depending on if-else from suffix option).
Framerate selection does not work consistently and needs some more work.
Does `ffmpeg` drop this option sometimes?