Added script to find the best exposure among many
I struggled to find the best image in a set of images taken before,
during and after the x-ray pulse.
Since this image should have a (much) higher mean than all the others in
the set, this script calculates the mean of all the images in a given
folder, looks for the maximum of this mean and tells the user about it.
It also plots said mean making it relatively quick to to trough a set of
folders each containing a set of images.