Reliably deleting and general cleanup
Files are only deleted if we have more than twenty images left.
Deletion happens a bit more clever.
Switching from ImageJ to Fiji at the end, since we need the
stack-functionality of it (average or sum images of stack).
Generally being more considerate with errors, i.e. when we cannot see an
image