Radiography-Detector finalized
Fully `pep8` and nearly full `pylint` conformity (I don't want to rename
all variables...)
Cleaning up description.
Cleaning up imports.
Bringing different options to the front of the file, so we don't have to
look for them.
Add lots of comments all around the place
Redo the final call to Fiji, so we can save the files as _average.tif
and _sum.tif in the end.