Discussing with Peter, correctly plotting MTF now
I had a quite long discussion with Peter concerning the MTF and
calculation of it, and think that now I have it calculated and shown
correctly (but very ugly).

Switching all plots to gridspec, which makes it easier to arrange plots
on the canvas.
The horizontal and vertical lines in the FFT are the MTF (horizontal and
vertical, respectively). We plot those and thus should be able to
distinguish between a bad and a good lens.
Adding some notes at the end.

TODO: Cleanup of the code, i.e. transfer some of the plotting logic into
own functions, so that we don't have to write the same thing over and
over. But this is something for another day...