Drawing is now (technically) correct.
Added lens-ellipse, even though we still operate in the thin-lens
regime.
Changed color of the lens representation to make it less "imposing".

Drawing a ellipsoid on top of the FOV to have a more "natural"
representation of the view of each CMOS.
Calculating the focal length via the distance from the CMOS to the lens,
so we have (theoretically) only one variable.
Split out second figure into subplots of top and side view.