"Modernized" GDX.ini file
We're now consistently using `clear` to get rid of the noise on the
command line.
Getting rid of redundant code in the sensor-naming block
Saving all images to consistent directories, based on `DevWare` and then
subdirectories with the name of the currently mounted sensor.
Give out more meaningful info for loading the images in Fiji