Added some logic in exposure time step selection
Unfortunately I'm not able to get a float range, so we have to round to
the next ms in the sensor exposure time.
Since we're anyways most probably have only larger exposure times, we
can live with this constraint.
Output is more nicely formatted.
According to the DevWare manual (http://is.gd/EtEwf8) we can have a
preset at the beginning which is exectued when the INI file is loaded.
Using that feature to de-duplicate some code.