Initial version of INI file for Ivans experiments
File is based on lots of stuff we have in [/aptina/GDX.ini].
At the moment we can set some states of the CMOS and get some parameters
for the file name.
Setting exposure time works (at least for AR0132, since the other eval
boards are out of the house at the moment).
Now going to introduce some logic about saving files as fast as
possible.