INI file tweaks for OS X
The Aptina folks provided me with an updated version of DevWare
(thanks!), where the `input` command also works on OS X.
Updating the INI file with some platform-if-calls to have it work in
both cases.
Currently, DevWare locks up while saving the images (on OS X), but does
work in the background.