Show only a subset of values
Maybe it would be better to use a masked array or a dictionary, but by
using the idea presented in http://stackoverflow.com/a/3555387/323100 we
can *easily* show only a subset of values, currently only the
experiments with the AppScinTech-szintillator.

Starting to optimize presentation with semi-transparent labels and the
like.
There's still *lots* of work in this departement :)

Also removing the /usr/bin/python to possibly help with Ivans homebrewn
python installation