Add routine to only redo evaluation when necessary
When the ID.resolution.png image is *not* here, we do the evaluation.
When it's there already, we present it to the user and let him decide to
do the evaluation or not.
Also only running tight_layout on Linux, since it's not really working
on OS X...
Adapting presentation