Adding script to analyze experiments
Script is based on AnalyzeGrid.py

Analysis will be done based on brightness of resulting image.
We cannot only read from the log file, since in the log file we only
specify the max/mean/STD of the best image.
We'll have to read in the images generated by
DetectWhichImageIsRadiography.py and redo the max/mean/std detection on
these files.