Correct calculation
Maïka notified me that it's unclear what eta stands for.
It was relatively easy to explain that this is the efficiency of x-ray
photon generation (as per section 3.1.5 of Krestel1990), but when going
through I found several errors (some with the help of Zhentian).
I didn't multiply eta with the atomic number of Tungsten and the mean
energy of the beam.
Additionally, I had some errors in the switching of units.
These errors have been fixed, now the calculation gives the same results
as Zhentians base document (for the case of 40kVp, 25 mA and 5 cm FOV at
1.2m distance)